# server/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: trades
    description: Manual trade journal operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List journal trades
      description: Returns manually recorded trades, newest first.
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: side
          in: query
          required: false
          schema:
            type: string
            enum: [buy, sell]
      responses:
        "200":
          description: Journal trades
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trade"
    post:
      operationId: createTrade
      tags: [trades]
      summary: Record a trade
      description: Adds a manual journal entry. This never places or modifies a broker order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      responses:
        "201":
          description: Created trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "400":
          description: Invalid trade data
  /trades/summary:
    get:
      operationId: getTradeSummary
      tags: [trades]
      summary: Get journal summary
      description: Returns aggregate counts and P&L for the current journal.
      responses:
        "200":
          description: Journal summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeSummary"
  /trades/{id}:
    patch:
      operationId: updateTrade
      tags: [trades]
      summary: Update a journal trade
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TradeInput"
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated trade
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trade"
        "400":
          description: Invalid trade data
        "404":
          description: Trade not found
    delete:
      operationId: deleteTrade
      tags: [trades]
      summary: Delete a journal trade
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Trade deleted
        "404":
          description: Trade not found
  /journal-notes:
    get:
      operationId: listJournalNotes
      summary: List dated notes and their screenshots
      responses:
        "200":
          description: Notes, newest journal date first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JournalNote"
    post:
      operationId: createJournalNote
      summary: Save a dated journal note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalNoteInput"
      responses:
        "201":
          description: Saved note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalNote"
        "400":
          description: Invalid note or trade links
  /journal-notes/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
    put:
      operationId: updateJournalNote
      summary: Replace a journal note and its screenshots
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalNoteInput"
      responses:
        "200":
          description: Updated note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JournalNote"
        "400":
          description: Invalid note or trade links
        "404":
          description: Note not found
    delete:
      operationId: deleteJournalNote
      summary: Delete a note and its screenshots without deleting trades
      responses:
        "204":
          description: Note deleted
        "404":
          description: Note not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Trade:
      type: object
      properties:
        id:
          type: integer
        symbol:
          type: string
          minLength: 1
        asset:
          type: string
          minLength: 1
        side:
          type: string
          enum: [buy, sell]
        lotSize:
          type: number
          exclusiveMinimum: 0
        entryAt:
          type: string
          format: date-time
        strategy:
          type: string
          minLength: 1
        regime:
          type: string
          minLength: 1
        pnl:
          type: number
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - symbol
        - asset
        - side
        - lotSize
        - entryAt
        - strategy
        - regime
        - pnl
        - notes
        - createdAt
        - updatedAt
    TradeInput:
      type: object
      properties:
        symbol:
          type: string
          minLength: 1
        asset:
          type: string
          minLength: 1
        side:
          type: string
          enum: [buy, sell]
        lotSize:
          type: number
          exclusiveMinimum: 0
        entryAt:
          type: string
          format: date-time
        strategy:
          type: string
          minLength: 1
        regime:
          type: string
          minLength: 1
        pnl:
          type: number
        notes:
          type: ["string", "null"]
      required:
        - symbol
        - asset
        - side
        - lotSize
        - entryAt
        - strategy
        - regime
        - pnl
    TradeSummary:
      type: object
      properties:
        totalTrades:
          type: integer
        buys:
          type: integer
        sells:
          type: integer
        totalPnl:
          type: number
        symbolsTraded:
          type: integer
        winningTrades:
          type: integer
        losingTrades:
          type: integer
      required:
        - totalTrades
        - buys
        - sells
        - totalPnl
        - symbolsTraded
        - winningTrades
        - losingTrades

    NoteAttachment:
      type: object
      required: [id, name, caption, dataUrl]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          minLength: 1
          maxLength: 255
        caption:
          type: string
          maxLength: 2000
        dataUrl:
          type: string
          maxLength: 6990600
          description: Base64 PNG, JPEG, WebP or GIF, at most 5 MiB decoded
    JournalNoteInput:
      type: object
      required: [date, timeZone, title, body, linkMode, selectedTradeIds, attachments]
      properties:
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          description: Calendar date in the saved timezone
        timeZone:
          type: string
          minLength: 1
          maxLength: 100
        title:
          type: string
          minLength: 1
          maxLength: 200
        body:
          type: string
          maxLength: 100000
        linkMode:
          type: string
          enum: [date, selected]
        selectedTradeIds:
          type: array
          maxItems: 1000
          uniqueItems: true
          items:
            type: integer
            minimum: 1
        attachments:
          type: array
          maxItems: 5
          items:
            $ref: "#/components/schemas/NoteAttachment"
    JournalNote:
      allOf:
        - $ref: "#/components/schemas/JournalNoteInput"
        - type: object
          required: [id, createdAt, updatedAt]
          properties:
            id:
              type: integer
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time